[EJBCLIENT-424] EJB remote invocation response payload contain ContextData not part of 'jboss.returned.keys'
When unmashalling an incoming EJB request, ContextData to preserve is collected
in a 'retainContextDataKeys' Set. These contain all keys provided by the client
as comma separated list for the key 'jboss.returned.keys', as well as internally
added keys. Before marshalling the response the attachments are reduced to the
keys contained in the 'retainContextDataKeys' Set.